+++ src/main/java/io/github/thebusybiscuit/slimefun4/api/ErrorReport.java
@@ -17,6 +17,7 @@ import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.plugin.Plugin;
 
+import io.github.thebusybiscuit.slimefun4.core.attributes.EnergyNetProvider;
 import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
 import me.mrCookieSlime.Slimefun.Objects.handlers.BlockTicker;
@@ -115,7 +116,7 @@ public class ErrorReport {
                 stream.println();
             }
 
-            if (item.getEnergyTicker() != null) {
+            if (item instanceof EnergyNetProvider) {
                 stream.println("Ticker-Info:");
                 stream.println("  Type: Indirect (Energy Network)");
                 stream.println();