+++ src/main/java/io/github/thebusybiscuit/slimefun4/api/exceptions/WrongItemStackException.java
@@ -30,7 +30,7 @@ public class WrongItemStackException extends RuntimeException {
      *            An error message to display
      */
     public WrongItemStackException(String message) {
-        super("You probably wanted alter a different ItemStack: " + message);
+        super("You probably wanted to alter a different ItemStack: " + message);
     }
 
 }