+++ src/main/java/me/mrCookieSlime/Slimefun/api/Slimefun.java
@@ -4,7 +4,6 @@ import java.util.Optional;
 import java.util.logging.Logger;
 
 import org.bukkit.Bukkit;
-import org.bukkit.NamespacedKey;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.scheduler.BukkitTask;
@@ -14,7 +13,6 @@ import io.github.thebusybiscuit.slimefun4.api.items.ItemState;
 import io.github.thebusybiscuit.slimefun4.api.player.PlayerProfile;
 import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
 import io.github.thebusybiscuit.slimefun4.implementation.items.VanillaItem;
-import me.mrCookieSlime.Slimefun.Objects.Research;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
 
 /**
@@ -30,48 +28,6 @@ public final class Slimefun {
 
     public static Logger getLogger() {
         return SlimefunPlugin.instance().getLogger();
-    }
-
-    /**
-     * Registers a research.
-     * 
-     * @deprecated The Research class was moved, this method is no longer valid. Please use
-     *             {@link io.github.thebusybiscuit.slimefun4.core.researching.Research#register()} instead.
-     * 
-     * @param research
-     *            The research
-     * @param items
-     *            The items
-     */
-    @Deprecated
-    public static void registerResearch(Research research, ItemStack... items) {
-        for (ItemStack item : items) {
-            research.addItems(SlimefunItem.getByItem(item));
-        }
-
-        research.register();
-    }
-
-    /**
-     * Registers a research.
-     * 
-     * @deprecated The Research class was moved, this method is no longer valid. Please use
-     *             {@link io.github.thebusybiscuit.slimefun4.core.researching.Research#register()} instead.
-     * 
-     * @param key
-     *            The key
-     * @param id
-     *            The id
-     * @param name
-     *            The name
-     * @param cost
-     *            The default cost
-     * @param items
-     *            The items
-     */
-    @Deprecated
-    public static void registerResearch(NamespacedKey key, int id, String name, int cost, ItemStack... items) {
-        registerResearch(new Research(key, id, name, cost), items);
     }
 
     /**